annotest_idl: split up be_init.cpp
